# chore(e2e): Use `wxt prepare` instead of `wxt build` when possible to speed up E2E tests