| Case | Before | After |
| --- | --- | --- |
| `pnpm test run` on Windows CI | 44s | 45s |
| `pnpm test:coverage` on Linux CI | 37s | 37s |

So no actual change to test performance